% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizing.R
\name{panel2plotsize}
\alias{panel2plotsize}
\title{Determine ggplot size from panel size}
\usage{
panel2plotsize(plot, panel_width, panel_height)
}
\arguments{
\item{plot}{A ggplot object or ggplotGrob}

\item{panel_width}{Desired panel width in mm}

\item{panel_height}{Desired panel height in mm}
}
\value{
Returns a list of computed dimensions
}
\description{
This function calculates the total size of a ggplot based on the desired
panel size.
}
